pyproject.toml
[project]
name = "rsrch-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.4.0",
    "ddgs>=9.10.0",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-google-genai>=4.1.3",
    "langchain-huggingface>=1.2.0",
    "marker-pdf",
    "psutil>=7.2.1",
    "pydantic>=2.12.5",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
]
